+++ js/utils/filters.js
@@ -1,3 +1,7 @@
+const filterDiscussed = document.querySelector('#filter-discussed');
+const filterRandom = document.querySelector('#filter-random');
+const filterDefault = document.querySelector('#filter-default');
+
 const compareComments = (pictureA, pictureB) => {
 
   if(pictureA.comments.length > pictureB.comments.length) {
@@ -11,5 +15,31 @@ const compareComments = (pictureA, pictureB) => {
 
 const mixPictures = () => Math.random() > 0.5 ? 1 : -1;
 
+const initPictureFilters = (picturesData, RERENDER_DELAY, renderPicturesList) => {
+  filterDiscussed.addEventListener('click', _.debounce(
+    () => {
+      const sortedPictureData = [...picturesData];
+      sortedPictureData.sort(compareComments);
+      renderPicturesList(sortedPictureData);
+    },
+    RERENDER_DELAY,
+  ));
 
-export {compareComments, mixPictures};
+  filterRandom.addEventListener('click', _.debounce(
+    () => {
+      const sortedPictureData = [...picturesData];
+      sortedPictureData.sort(mixPictures);
+      renderPicturesList(sortedPictureData.slice(0,10));
+    },
+    RERENDER_DELAY,
+  ));
+
+  filterDefault.addEventListener('click', _.debounce(
+    () => {
+      renderPicturesList(picturesData);
+    },
+    RERENDER_DELAY,
+  ));
+};
+
+export {initPictureFilters};
